requirements.txt
streamlit>=1.29.0
google-generativeai>=0.3.2
PyPDF2>=3.0.1
python-docx>=0.8.11
faiss-cpu>=1.7.4
numpy>=1.24.3
pandas>=2.0.3
python-dotenv>=1.0.0
tiktoken>=0.5.1
